Split Bus into interface and implementation
The implementation is called Bus, the implementation is called
BusImpl.

The primary purpose of this change is to facilitate connector tests in
which the bus is mocked.
